requirements.txt
pandas==2.3.3
scikit-learn==1.7.2
numpy==2.2.6
wandb==0.22.1
pyyaml==6.0.3
fastapi==0.119.0
uvicorn==0.38.0
streamlit==1.50.0
pytest==8.4.2